- Add LeaveRequestHistory entity and tracking for edits/updates
- Implement LeaveBalanceRepository for CRUD and auto-creation
- Add approval/rejection commands with email notifications
- Introduce HTML email templates for leave workflows
- Add dummy JWT token generation endpoint for testing
- Change dayAmount to decimal; remove isEdit/initialRequestId from model
- Update LeaveManagementController with new endpoints
- Register new repositories in DI container
- Improve exception middleware for dev debugging
- Relax employee email validation rules
- Update DbContext/interfaces for new leave tables
- General cleanup and improved logging throughout

+++ HRManagement.Api/Domain/Models/Tables/LeaveManagementModel/LeaveRequest/CreateLeaveRequestDto.cs
@@ -6,7 +6,7 @@
         public int SupervisorId { get; set; }
         public string LeaveDescription { get; set; }
         public DateTime leaveStartDate { get; set; }
-        public int DayAmount { get; set; }
+        public decimal DayAmount { get; set; }
         public int LeaveType { get; set; }
         public string[] AttachmentPath { get; set; }
